+++ src/gfx/bmp_to_png.cpp
@@ -6,8 +6,8 @@
 const int ltwid = 256;
 const int lthei = 256;
 #ifndef WIN32
-typedef unsigned long DWORD;
-typedef long  LONG;
+typedef unsigned int DWORD;
+typedef int  LONG;
 typedef unsigned short WORD;
 typedef unsigned char BYTE;
 /**
